+++ web/src/pages/ResetPasswordPage/ResetPasswordPage.tsx
@@ -1,0 +1,120 @@
+import { useEffect, useRef, useState } from 'react'
+
+import { useAuth } from '@redwoodjs/auth'
+import {
+  Form,
+  Label,
+  PasswordField,
+  Submit,
+  FieldError,
+} from '@redwoodjs/forms'
+import { navigate, routes } from '@redwoodjs/router'
+import { MetaTags } from '@redwoodjs/web'
+import { toast, Toaster } from '@redwoodjs/web/toast'
+
+const ResetPasswordPage = ({ resetToken }: { resetToken: string }) => {
+  const { isAuthenticated, reauthenticate, validateResetToken, resetPassword } =
+    useAuth()
+  const [enabled, setEnabled] = useState(true)
+
+  useEffect(() => {
+    if (isAuthenticated) {
+      navigate(routes.home())
+    }
+  }, [isAuthenticated])
+
+  useEffect(() => {
+    const validateToken = async () => {
+      const response = await validateResetToken(resetToken)
+      if (response.error) {
+        setEnabled(false)
+        toast.error(response.error)
+      } else {
+        setEnabled(true)
+      }
+    }
+    validateToken()
+  }, [])
+
+  const passwordRef = useRef<HTMLInputElement>(null)
+  useEffect(() => {
+    passwordRef.current?.focus()
+  }, [])
+
+  const onSubmit = async (data: Record<string, string>) => {
+    const response = await resetPassword({
+      resetToken,
+      password: data.password,
+    })
+
+    if (response.error) {
+      toast.error(response.error)
+    } else {
+      toast.success('Password changed!')
+      await reauthenticate()
+      navigate(routes.login())
+    }
+  }
+
+  return (
+    <>
+      <MetaTags title="Reset Password" />
+
+      <main className="main">
+        <Toaster toastOptions={{ className: 'toast', duration: 6000 }} />
+        <div className="scaffold login-container">
+          <div className="segment">
+            <header className="segment-header">
+              <h2 className="heading heading-secondary">
+                Reset Password
+              </h2>
+            </header>
+
+            <div className="segment-main">
+              <div className="form-wrapper">
+                <Form onSubmit={onSubmit} className="form-wrapper">
+                  <div className="text-left">
+                    <Label
+                      name="password"
+                      className="label"
+                      errorClassName="label label-error"
+                    >
+                      New Password
+                    </Label>
+                    <PasswordField
+                      name="password"
+                      autoComplete="new-password"
+                      className="input"
+                      errorClassName="input input-error"
+                      disabled={!enabled}
+                      ref={passwordRef}
+                      validation={{
+                        required: {
+                          value: true,
+                          message: 'New Password is required',
+                        },
+                      }}
+                    />
+
+                    <FieldError name="password" className="field-error" />
+                  </div>
+
+                  <div className="button-group">
+                    <Submit
+                      className="button button-blue"
+                      disabled={!enabled}
+                    >
+                      Submit
+                    </Submit>
+                  </div>
+                </Form>
+              </div>
+            </div>
+          </div>
+        </div>
+      </main>
+    </>
+  )
+}
+
+export default ResetPasswordPage