+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## 2023.01.14
 
+- updates gitignore for seeds
+- tests initial db seed migration of basic users
+- updates db schema
+- initialize dbauth + ui setup
+- sample toggle background button component (archived)
+- new loginoroutlink component
+- new userdetails component for quick view of currently logged in user info
+- minor css updates
+- prep for live push test
+
 - initial commit
 - add css reset
 - starts devnotes & changelog